chore: reorganize React architecture views
