+++ src/components/index.ts
@@ -1,8 +1,8 @@
-export * from "./AnimateIn/AnimateIn";
-export * from "./Layout/Layout";
-export * from "./Footer/Footer";
-export * from "./Header/Header";
-export * from "./Meta/Meta";
-export * from "./Preloader/Preloader";
-export * from "./ScrollTop/ScrollTop";
-export * from "./VideoCard/VideoCard";
+export * from "./AnimateIn";
+export * from "./Layout";
+export * from "./Footer";
+export * from "./Header";
+export * from "./Meta";
+export * from "./Preloader";
+export * from "./ScrollTop";
+export * from "./VideoCard";
